Fix #453: Show program end time in [GraphicalEPGPIG]
[GraphicalEPGPIG]
Add program end time to the top line of the information display for the
event in focus.

Give the title line display the same left alignment as the description.

Adjust the location and width of the show title to accommodate.

Adjust the widths of the description and bouquet name fields to have the
same right edge as the title line.