Fix bug #532: [skin] Fill anomalies in MoviePlayer Infobar
Remove the plain colour fill and replace it by extending the "bottom"
fill navy_fondo-fs8.png up to cover the area previously covered by
the colour fill.

Change height of MoviePlayer infobar screen to match the actual
height of the infobar.

Remove the alphatest="on" attribute from the navy_fondo-fs8.png
pixmap in the MoviePlayer infobar screen to match the pixmap
attributes in InfoBarCommon.

Change the comments on the fill area to reflect what they actually
are.