[skin_infobars] Adjust InfoBarClock for enhanced time formats
Use enhanced time formats in the InfoBarClock panel.

Adjust the width of the panel for the added width of the AM/PM times.

Adjust the width of panel uses to match the panel.

Move the separator between the InfoBarClock and the rest of
the InfoBarCommon for the live TV infobar to accommodate the
extra width of the panel.