[skin] Adjust InfoBarEventView start/end times for enhanced time formats
Un-comment the Beyonwiz-specific version of InfoBarEventView so it
can be modified.

Replace time/duration widgets with widgets using Source/Converter/Renderer
and enhanced time formats for start/end times and duration.

Adjust the location and sizes of the begin/end times and duration
widgets to accommodate AM/PM times, to allow the display of descenders
in the month nams and to align the baselines of the title, start/end
times and duration firlds.

Add empty widget definitions for all unused widgets in the class,
including the now unused "datetime" and "duration" widgets.