[skin_epg] Adjust QuickEPG for enhanced time formats
[QuickEPG]

Skin for the infobar single EPG.

Copy QuickEPG from skin_defaults so that it can be adjusted.

Use the coloured button hints and text alignment from the Beyonwiz
skin so that it matches other Beyonwiz "cramped space" screens
better.

For the other icons, add a skin_defaults prefix to their paths so
that the icons that match the original screen are used.

Change date and time formats from fixed formats to standard time
format names.  A single font size is now used for the whole time,
rather than a smaller font for seconds.

The Mixed time format is used so that in am/pm formats, the "am/pm"
replaces the seconds field that is displayed in the 24-hour formats.
This means that the same font size as the original can be kept.

The date format has been changed so that it doesn't have the year.
This means that the longer ShortDate format fits in the "recess"
in the underlying pixmap without changing the font size.

Adjust the width of the date and time fields so that they have a
balanced amount of space between the ends of the field and the
"recess".

Adjust the height of the time field so that the descenders in the
"ap/pm" field are displayed. Adjust the vertical alignment of the
fields so that the date/time block is vcentred verticaly in the
"recess"

Adhust the height of the FullDescription field so that it is an
even multiple of the font lead.

Add skin_defaults as a prefix to most Adjust the vertical positions
and heights of the current date and time widgets so that vertically
descenders in the am/pm time format don't overlap with the date
text.