[displaytext] Split parameterisation of enhanced time formats
Enable AllowUserDatesAndTimes in the scren skins and enable
AllowUserDatesAndTimesDisplay only in the displaytext front panel
skin (for the T2 & T3).

This means that the setting MENU>Setup>System>Front panel settings>Time
style will be available on only the T2 & T3. It should really only
be enabled on the T3, but it is enabled on the T3 because the two
models share front panel skins.