[AutoTimer] Move setup description to standard setup screen locations
The SetupCommon panel can't be used because the AutoTimer uses
a source/render widgit rether than of a plain widget as in SetupCommon.

In preparation for fixing #550: [easy-skin-aus-hd skin] Not enough
space for some AutoTimer edit settings.

Although Full-Metal-Wizard isn't affected by #550, make the same changes
as in easy-skin-aus-hd so that the AutoTimer screens better match the
standard setup screens.