Fix bug #614: [skins] Morning end times sometimes not shown in infobar for hh.mm.ssAM/PM time
This change isn't strictly needed for the Full-Metal-Wizard skin,
but made so that the parts of the skin that were the same as for
the easy-skin-aus-hd skin remain the same.

Give the now/next start/end fields in the infobar a little more
space, and set noWrap so that if the text is to large, it is truncated
at the field edge rather than being wrapped out of the containing
box.