Fix issue #623: Rewrite infobar Simple (single channel) EPG to match infobar and add information to match other EPG event info
Re-write the infobar Simple EPG (QuickEPG) so that it better matches
the style of the skin and to use components from the infobar. Add
more use of panels to construct the infobar so that the panels can
be re-used in QuickEPG.

Add taller version of background pixmap than used for normal infobar.