Fix issue #631: [Full-Metal-Wizard] menu/mainmenu_* pixmaps have inappropriate position and size
Change the screen sizes of the widgets and ePixmaps containing
menu/mainmenu_* pixmaps in Full-Metal-Wizard, re-centre them and
move them down a little.

Remove Full-Metal-Wizard/ and easy-skin-aus-hd/ out-of-skin path
prefixes on the affected pixmap paths. Removing the easy-skin-aus-hd/
mens that those pixmaps change to ones with Full-Metal-Wizard style.

Add "plugin" icon to the Download plugins and Remove plugins screens
(PluginDownloadBrowser), and move and resize the "text" widget to
accommodate the icon.