Fix bug #648: Poor text alignment in skin button hints
Use noWrap on all coloured button hint label widgets.

Widen the coloured area of all the coloured button hint pixmaps
so that they are the full 200px widdth, including making the
blue pixmap coloured area the same width as in the other buttons.

Adjust the width and position of the button hint label widgets
so that they are enclosed in the coloured area of the pixmap
and so the text will be correctly centred on the coloured area.