Fix issue #673: Tidy up FileCommander list headers
Use two widgets for the header entries, one a name widget
for the file path and the other a source with a
List/TemplatedMultiContent/Listbox chain to display the
file status.

Display the symbolic file mode, the scaled file size
and the modification time in the file mode.

Put the File Commander file lists skin into a panel
so that the panel can be used in both FileCommanderScreen and
FileCommanderScreenFileSelect screen skins.