Allow up to 15 tuner state indicators & customise transponder info
[skin_infobars.xml]

Change the display of the tuner state indicators from using 4
individual tuner tests to using the FrontendInfo STRING conversion
to display the tuner info as a string. There is space in its widget
for the display of 15 indicators.

Use a custom PliExtraInfo "User" conversion to display provider
name, transponder channel number, transponder frequency in MHz and
transponder bandwidth in MHz, rather than using the built-in
"TransponderInfo2line" conversion.

This removes some clutter, gives a more readable frequency and
reduces the information to a single line to accommodate the new
tuner state display.

[skin.xml]

Add FrontendInfoColors parameter to give the expected green/red/darkergrey
indicator colours instead of the code defaults of green/yellow/grey.