Remove irrelevant service video icons from infobar
[skin_infobars.xml]

Remove the service HbbTV, Video Text or 16:9 indicators as they're
not of any real use.

Align the right edge of the icon bar with the right edge of the
items below it.

Keep the "update available" icon in the same position with its left
edge aligned with left edge of the items below it, as it's not
really associated with the other icons in the row.